% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_info.R
\name{mef_info}
\alias{mef_info}
\title{Return a structure containing MEF header information.
Based on Matlab code "read_mef_discont".}
\usage{
mef_info(args)
}
\arguments{
\item{filename}{String: The complete path to a .mef file}

\item{password}{String: The public password for the MEF file.}
}
\description{
Return a structure containing MEF header information.
Based on Matlab code "read_mef_discont".
}
