% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEFcont.R
\name{MEFcont}
\alias{MEFcont}
\title{Creates an iterator of contiguous MEF sequences.}
\usage{
MEFcont(filename, password, ...)
}
\arguments{
\item{filename}{String: The complete path to a .mef file}

\item{password}{String: The public password for the MEF file.}

\item{block0}{Integer: Start block number.}

\item{block1}{Integer: Stop block number.}

\item{time0}{Integer: Start time value (in microseconds)}

\item{time1}{Integer: Stop time value (in microseconds)}

\item{step}{Integer: Time step (in seconds) for each iteration.}
}
\value{
A data iterator for MEF files.
}
\description{
Creates an iterator of contiguous MEF sequences.
}
\examples{
\dontrun{
  data_iter <- MEFiter( filename, password, time0=1.00E15, time1=1.01E15, step=10 )
  data <- nextElem( data_iter )
}
}
