% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEFiter.R
\name{MEFiter}
\alias{MEFiter}
\title{Create a data iterator for MEF files}
\usage{
MEFiter(filename, password, ...)
}
\arguments{
\item{filename}{String: The complete path to a .mef file}

\item{password}{String: The public password for the MEF file.}

\item{block0}{Integer: Start block number.}

\item{block1}{Integer: Stop block number.}

\item{time0}{Integer: Start time value (in microseconds)}

\item{time1}{Integer: Stop time value (in microseconds)}

\item{stepSize}{Integer: Timestep size (in seconds) for each iteration.}
}
\value{
A data iterator for MEF files.
}
\description{
Create a data iterator for MEF files
}
\examples{
\dontrun{
  data_iter <- MEFiter( filename, password, time0=1.00E15, time1=1.01E15, stepSize=10 )
  data <- nextElem( data_iter )
}
}
