% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLiter.R
\name{SQLiter}
\alias{SQLiter}
\title{Return an iterator of database records.}
\usage{
SQLiter(db, query)
}
\arguments{
\item{db}{MySQL database connection:}

\item{query}{String: "select ..." - it is assumed that this will always be a "select" query.}
}
\description{
Return an iterator of database records.
}
